maintenance + basic 2D renderer

+++ Sapphire/Source/Sapphire/Time.cpp
@@ -7,7 +7,7 @@ namespace sph
 {
 	void Time::Update()
 	{
-		GameTime = glfwGetTime();
+		GameTime = (float)glfwGetTime();
 
 		UnscaledDeltaTime = GameTime - LastTime;
 		DeltaTime = UnscaledDeltaTime * TimeScale;
